NN constructor to if-based build
Changed the DNN layer builder in the actornetwork and criticnetwork classes to a build based on if statements, it only works for a maximum of 3 layers now but that's all that's needed for the experiments. It creates the layers directly as attributes of the NN classes and thus it can run on the GPU (and stays closer to the orgininal code). Last commit before validation run with standard settings.